where_watch/requirements.txt
aerich==0.6.2
aioredis==2.0.1
aiosqlite==0.17.0
amqp==5.0.9
anyio==3.5.0
asgiref==3.4.1
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
beautifulsoup4==4.10.0
bidict==0.21.4
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
databases==0.5.4
dataclasses-json==0.5.6
ddlparse==1.10.0
Deprecated==1.2.13
dictdiffer==0.9.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.72.0
fastapi-cache2==0.1.8
fastapi-socketio==0.0.9
greenlet==1.1.2
h11==0.12.0
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
httpcore==0.14.4
httptools==0.2.0
httpx==0.21.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iso8601==0.1.16
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.3
lxml==4.7.1
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
orjson==3.6.6
packaging==21.3
parsel==1.6.0
pendulum==2.1.2
pep8==1.7.1
priority==1.3.0
prompt-toolkit==3.0.24
Protego==0.1.16
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.6
pypika-tortoise==0.1.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.1
python-multipart==0.0.5
python-socketio==5.5.1
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
queuelib==1.6.2
redis==4.1.1
requests==2.27.1
rfc3986==1.5.0
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
starlette==0.17.1
toml==0.10.2
tomlkit==0.9.0
tortoise-orm==0.18.1
Twisted==21.7.0
typing-inspect==0.7.1
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
vine==5.0.0
w3lib==1.22.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
wrapt==1.13.3
zope.interface==5.4.0
